requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
supabase>=2.0.0
asyncpg>=0.29.0

# AI/ML
openai>=1.3.0
anthropic>=0.7.0

# Parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Note: httpx already listed under Testing
